Move controls to Controls namespace
