+++ Fluentver/Controls/GlyphButton.xaml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<UserControl
+    x:Class="Fluentver.Controls.GlyphButton"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Fluentver.Controls"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <Button x:Name="button" Click="Button_Click">
+        <StackPanel Spacing="8" Orientation="Horizontal">
+            <FontIcon Glyph="{x:Bind Glyph}" FontSize="16" />
+            <TextBlock Text="{x:Bind Text}" />
+        </StackPanel>
+    </Button>
+</UserControl>
