+++ Fluentver/Controls/UserEntry.xaml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<UserControl
+    x:Class="Fluentver.Controls.UserEntry"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Fluentver.Controls"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <StackPanel Spacing="8" Orientation="Horizontal" Padding="0,12,5,12">
+        <PersonPicture ProfilePicture="{x:Bind ProfilePicture}" Height="40" Width="40" />
+        <StackPanel VerticalAlignment="Center">
+            <TextBlock Text="{x:Bind DisplayName}" IsTextSelectionEnabled="True" />
+            <TextBlock Text="{x:Bind AccountName}" IsTextSelectionEnabled="True" Foreground="{ThemeResource TextFillColorSecondary}" />
+        </StackPanel>
+    </StackPanel>
+</UserControl>
